refactor(test): move mutation tests to git_interop module
Moved tests from standalone file to git_interop.rs test module for several benefits:
- Can now test private functions directly (e.g., new_symbolic_write_ref)
- Better organization - tests live with the code they test
- Removed markdown analysis document (details in PR description instead)

Tests now directly verify:
- new_symbolic_write_ref() returns valid, non-empty ref names with hex suffixes
- add_note_line_to_head() end-to-end functionality
- walk_commits() behavior in shallow vs normal repos

All tests target missed mutants identified in weekly mutation testing analysis.
