--- git_perf/tests/git_interop_mutation_tests.rs
@@ -1,181 +1,0 @@
-// Tests specifically targeting missed mutants found in mutation testing analysis
-// See mutation-testing-analysis.md for details
-//
-// Note: Some of the missed mutants are in private functions that cannot be tested directly.
-// These tests target the public APIs and integration scenarios that exercise the mutated code paths.
-
-use std::env::{self, set_current_dir};
-use std::path::Path;
-use std::process::Command;
-use tempfile::{tempdir, TempDir};
-
-fn run_git_command(args: &[&str], dir: &Path) -> std::io::Result<std::process::Output> {
-    Command::new("git")
-        .args(args)
-        .envs([
-            ("GIT_CONFIG_NOSYSTEM", "true"),
-            ("GIT_CONFIG_GLOBAL", "/dev/null"),
-            ("GIT_AUTHOR_NAME", "testuser"),
-            ("GIT_AUTHOR_EMAIL", "testuser@example.com"),
-            ("GIT_COMMITTER_NAME", "testuser"),
-            ("GIT_COMMITTER_EMAIL", "testuser@example.com"),
-        ])
-        .current_dir(dir)
-        .output()
-}
-
-fn init_repo(dir: &Path) {
-    run_git_command(&["init", "--initial-branch", "master"], dir).unwrap();
-    run_git_command(&["commit", "--allow-empty", "-m", "Initial commit"], dir).unwrap();
-}
-
-fn dir_with_repo() -> TempDir {
-    let tempdir = tempdir().unwrap();
-    init_repo(tempdir.path());
-    tempdir
-}
-
-fn hermetic_git_env() {
-    env::set_var("GIT_CONFIG_NOSYSTEM", "true");
-    env::set_var("GIT_CONFIG_GLOBAL", "/dev/null");
-    env::set_var("GIT_AUTHOR_NAME", "testuser");
-    env::set_var("GIT_AUTHOR_EMAIL", "testuser@example.com");
-    env::set_var("GIT_COMMITTER_NAME", "testuser");
-    env::set_var("GIT_COMMITTER_EMAIL", "testuser@example.com");
-}
-
-/// Test that notes can be added successfully via git_interop API
-/// Indirectly tests new_symbolic_write_ref() which had missed mutants:
-/// - Could return Ok(String::new()) - empty string
-/// - Could return Ok("xyzzy".into()) - arbitrary invalid string
-///
-/// The function new_symbolic_write_ref() is private but called by ensure_symbolic_write_ref_exists(),
-/// which is called by add_note_line_to_head(). If the ref name returned by new_symbolic_write_ref()
-/// is empty ("") or invalid ("xyzzy"), the git operations will fail.
-///
-/// This test verifies the operation succeeds and can retrieve the note, which proves:
-/// 1. The symbolic ref name was not empty
-/// 2. The symbolic ref name was not an arbitrary invalid string
-/// 3. The ref was properly formatted for git operations
-#[test]
-fn test_symbolic_write_ref_creates_valid_reference() {
-    let tempdir = dir_with_repo();
-    set_current_dir(tempdir.path()).expect("Failed to change dir");
-    hermetic_git_env();
-
-    // Add first note - this calls ensure_symbolic_write_ref_exists -> new_symbolic_write_ref
-    // If new_symbolic_write_ref returns empty or invalid, this will fail
-    let result = git_perf::git::git_interop::add_note_line_to_head("test: 100");
-    assert!(
-        result.is_ok(),
-        "Should add note (requires valid ref from new_symbolic_write_ref): {:?}",
-        result
-    );
-
-    // Add second note to ensure ref operations continue to work
-    let result2 = git_perf::git::git_interop::add_note_line_to_head("test: 200");
-    assert!(result2.is_ok(), "Should add second note: {:?}", result2);
-
-    // Verify notes were actually added by walking commits
-    let commits = git_perf::git::git_interop::walk_commits(10);
-    assert!(commits.is_ok(), "Should walk commits: {:?}", commits);
-
-    let commits = commits.unwrap();
-    assert!(!commits.is_empty(), "Should have commits");
-
-    // Check that HEAD commit has notes
-    let (_, notes) = &commits[0];
-    assert!(!notes.is_empty(), "HEAD should have notes");
-    assert!(
-        notes.iter().any(|n| n.contains("test:")),
-        "Notes should contain our test data"
-    );
-}
-
-/// Test walk_commits with shallow repository containing multiple grafted commits
-/// Targets missed mutant at line 725: detected_shallow |= vs ^=
-/// The XOR operator would toggle instead of OR, failing with multiple grafts
-#[test]
-fn test_walk_commits_shallow_repo_detection() {
-    let tempdir = dir_with_repo();
-    hermetic_git_env();
-
-    // Create multiple commits
-    set_current_dir(tempdir.path()).unwrap();
-    for i in 2..=5 {
-        run_git_command(
-            &["commit", "--allow-empty", "-m", &format!("Commit {}", i)],
-            tempdir.path(),
-        )
-        .unwrap();
-    }
-
-    // Create a shallow clone (depth 2) which will have grafted commits
-    let shallow_dir = tempdir.path().join("shallow");
-    let output = Command::new("git")
-        .args(&[
-            "clone",
-            "--depth",
-            "2",
-            tempdir.path().to_str().unwrap(),
-            shallow_dir.to_str().unwrap(),
-        ])
-        .output()
-        .unwrap();
-
-    assert!(
-        output.status.success(),
-        "Shallow clone failed: {}",
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    set_current_dir(&shallow_dir).unwrap();
-    hermetic_git_env();
-
-    // Add a note to enable walk_commits
-    git_perf::git::git_interop::add_note_line_to_head("test: 100").expect("Should add note");
-
-    // Walk commits - should detect as shallow
-    let result = git_perf::git::git_interop::walk_commits(10);
-    assert!(result.is_ok(), "walk_commits should succeed: {:?}", result);
-
-    let commits = result.unwrap();
-
-    // In a shallow repo, git log --boundary shows grafted markers
-    // The |= operator correctly sets detected_shallow to true
-    // The ^= mutant would toggle the flag, potentially giving wrong result
-
-    // Verify we got commits (the function works)
-    assert!(
-        !commits.is_empty(),
-        "Should have found commits in shallow repo"
-    );
-}
-
-/// Test walk_commits correctly identifies normal (non-shallow) repos
-#[test]
-fn test_walk_commits_normal_repo_not_shallow() {
-    let tempdir = dir_with_repo();
-    set_current_dir(tempdir.path()).unwrap();
-    hermetic_git_env();
-
-    // Create a few commits
-    for i in 2..=3 {
-        run_git_command(
-            &["commit", "--allow-empty", "-m", &format!("Commit {}", i)],
-            tempdir.path(),
-        )
-        .unwrap();
-    }
-
-    // Add a note to enable walk_commits
-    git_perf::git::git_interop::add_note_line_to_head("test: 100").expect("Should add note");
-
-    let result = git_perf::git::git_interop::walk_commits(10);
-    assert!(result.is_ok(), "walk_commits should succeed");
-
-    let commits = result.unwrap();
-
-    // Should have commits
-    assert!(!commits.is_empty(), "Should have found commits");
-}
